Refactoring of TopThreeCardsViewHolder and FutureCard
