+++ app/src/main/java/com/example/se2_exploding_kittens/TopThreeCardsViewHolder.java
@@ -1,22 +1,20 @@
 package com.example.se2_exploding_kittens;
 
 import android.os.Handler;
-import android.os.Looper;
 import android.view.View;
 import android.widget.ImageView;
+import android.widget.PopupWindow;
 import android.widget.TextView;
 
 import androidx.recyclerview.widget.RecyclerView;
 
 public class TopThreeCardsViewHolder extends RecyclerView.ViewHolder {
-    private ImageView cardOne;
-    private ImageView cardTwo;
-    private ImageView cardThree;
-    private TextView textCardOne;
-    private TextView textCardTwo;
-    private TextView textCardThree;
-
-    private Handler handler;
+    private final ImageView cardOne;
+    private final ImageView cardTwo;
+    private final ImageView cardThree;
+    private final TextView textCardOne;
+    private final TextView textCardTwo;
+    private final TextView textCardThree;
 
     public TopThreeCardsViewHolder(View itemView) {
         super(itemView);
@@ -26,7 +24,6 @@ public class TopThreeCardsViewHolder extends RecyclerView.ViewHolder {
         textCardOne = itemView.findViewById(R.id.textViewFirstCard);
         textCardTwo = itemView.findViewById(R.id.textViewSecondCard);
         textCardThree = itemView.findViewById(R.id.textViewThirdCard);
-        handler = new Handler(Looper.getMainLooper());
     }
 
     public void bindData(int firstCard, int secondCard, int thirdCard) {
@@ -53,5 +50,33 @@ public class TopThreeCardsViewHolder extends RecyclerView.ViewHolder {
             cardThree.setVisibility(View.INVISIBLE);
             textCardThree.setVisibility(View.INVISIBLE);
         }
+    }
+
+    public void run(Handler handler, PopupWindow popupWindow, TextView timerTextView) {
+        // Delay between each update in milliseconds
+        final int delay = 1000;
+        handler.postDelayed(new Runnable() {
+            // Total time in milliseconds
+            int remainingTime = 5000;
+
+            @Override
+            public void run() {
+                if (popupWindow.isShowing()) {
+                    // Subtract the delay from the remaining time
+                    remainingTime -= delay;
+                    // Calculate seconds
+                    int seconds = remainingTime / 1000;
+
+                    if (seconds > 0) {
+                        timerTextView.setText(String.valueOf(seconds));
+                        // Schedule the next update
+                        handler.postDelayed(this, delay);
+                    } else {
+                        // Dismiss the PopupWindow after 5000ms
+                        popupWindow.dismiss();
+                    }
+                }
+            }
+        }, delay);
     }
 }
